refactor: 파일 경로 이동 & DashBoard로 이름 변경
